+++ pycryptobot.py
@@ -65,6 +65,11 @@ def executeJob(
 ):
     """Trading bot job which runs at a scheduled interval"""
 
+    if app.isLive():
+        state.account.mode = "live"
+    else:
+        state.account.mode = "test"
+
     # This is used to control some API calls when using websockets
     last_api_call_datetime = datetime.now() - _state.last_api_call_datetime
     if last_api_call_datetime.seconds > 60:
@@ -97,7 +102,21 @@ def executeJob(
 
         if controlstatus == "exit":
             _app.notifyTelegram(f"{_app.getMarket()} bot is stopping")
+            telegram_bot.removeactivebot()
             sys.exit(0)
+
+        if controlstatus == "reload":
+            _app.read_config(_app.getExchange())
+            if _app.enableWebsocket():
+                _websocket.close()
+                _websocket = BWebSocketClient([app.getMarket()], app.getGranularity())
+                _websocket.start()
+            _app.setGranularity(_app.getGranularity())
+            list(map(s.cancel, s.queue))
+            s.enter(5, 1, executeJob, (sc, _app, _state, _technical_analysis, _websocket))
+            # _app.read_config(_app.getExchange())
+            telegram_bot.updatebotstatus("active")
+
 
     # reset _websocket every 23 hours if applicable
     if _app.enableWebsocket() and not _app.isSimulation():
@@ -1318,6 +1337,8 @@ def executeJob(
                         ignore_index=True,
                     )
 
+                    state.in_open_trade = True
+
                     state.last_api_call_datetime -= timedelta(seconds=60)
 
                 if _app.shouldSaveGraphs():
@@ -1548,6 +1569,7 @@ def executeJob(
                         },
                         ignore_index=True,
                     )
+                    state.in_open_trade = False
                     state.last_api_call_datetime -= timedelta(seconds=60)
                 if _app.shouldSaveGraphs():
                     tradinggraphs = TradingGraphs(_technical_analysis)
@@ -1807,7 +1829,7 @@ def executeJob(
                     str(_truncate(margin, 4) + "%") if _state.in_open_trade == True else " ",
                     str(_truncate(profit, 2)) if _state.in_open_trade == True else " ",
                     price,
-                    change_pcnt_high,
+                    change_pcnt_high
                 )
             
             # Update the watchdog_ping