+++ telegram_bot.py
@@ -40,6 +40,7 @@ from models.telegram import (
     TelegramHandler,
     TelegramActions,
     ConfigEditor,
+    SettingsEditor
 )
 
 scannerSchedule = BackgroundScheduler(timezone="UTC")
@@ -56,9 +57,9 @@ CHOOSING, TYPING_REPLY = range(2)
 EXCHANGE, MARKET, ANYOVERRIDES, OVERRIDES, SAVE, START = range(6)
 EXCEPT_EXCHANGE, EXCEPT_MARKET = range(2)
 
-reply_keyboard = [["Coinbase Pro", "Binance", "Kucoin"]]
+replykeyboard = [["Coinbase Pro", "Binance", "Kucoin"]]
 
-markup = ReplyKeyboardMarkup(reply_keyboard, one_time_keyboard=True)
+markup = ReplyKeyboardMarkup(replykeyboard, one_time_keyboard=True)
 
 
 class TelegramBotBase:
@@ -74,7 +75,7 @@ class TelegramBotBase:
     handler = None
     editor = None
 
-    def _checkifallowed(self, userid, update) -> bool:
+    def _check_if_allowed(self, userid, update) -> bool:
         if str(userid) != self.userid:
             update.message.reply_text("<b>Not authorised!</b>", parse_mode="HTML")
             return False
@@ -123,45 +124,45 @@ class TelegramBot(TelegramBotBase):
         self.token = self.config["telegram"]["token"]
         self.userid = self.config["telegram"]["user_id"]
 
-        # Config section for bot pair scanner
-        self.atr72pcnt = 2.0
-        self.enableleverage = False
-        self.use_default_scanner = 1
-        self.maxbotcount = 0
-        self.autoscandelay = 0
-        self.enable_buy_next = True
-        self.autostart = False
-        if "scanner" in self.config:
-            self.atr72pcnt = (
-                self.config["scanner"]["atr72_pcnt"]
-                if "atr72_pcnt" in self.config["scanner"]
-                else self.atr72pcnt
-            )
-            self.enableleverage = (
-                self.config["scanner"]["enableleverage"]
-                if "enableleverage" in self.config["scanner"]
-                else self.enableleverage
-            )
-            self.use_default_scanner = (
-                self.config["scanner"]["use_default_scanner"]
-                if "use_default_scanner" in self.config["scanner"]
-                else self.use_default_scanner
-            )
-            self.maxbotcount = (
-                self.config["scanner"]["maxbotcount"]
-                if "maxbotcount" in self.config["scanner"]
-                else self.maxbotcount
-            )
-            self.autoscandelay = (
-                self.config["scanner"]["autoscandelay"]
-                if "autoscandelay" in self.config["scanner"]
-                else 0
-            )
-            self.enable_buy_next = (
-                self.config["scanner"]["enable_buy_next"]
-                if "enable_buy_next" in self.config["scanner"]
-                else True
-            )
+        # # Config section for bot pair scanner
+        # self.atr72pcnt = 2.0
+        # self.enableleverage = False
+        # self.use_default_scanner = 1
+        # self.maxbotcount = 0
+        # self.autoscandelay = 0
+        # self.enable_buy_next = True
+        # self.autostart = False
+        # if "scanner" in self.config:
+        #     self.atr72pcnt = (
+        #         self.config["scanner"]["atr72_pcnt"]
+        #         if "atr72_pcnt" in self.config["scanner"]
+        #         else self.atr72pcnt
+        #     )
+        #     self.enableleverage = (
+        #         self.config["scanner"]["enableleverage"]
+        #         if "enableleverage" in self.config["scanner"]
+        #         else self.enableleverage
+        #     )
+        #     self.use_default_scanner = (
+        #         self.config["scanner"]["use_default_scanner"]
+        #         if "use_default_scanner" in self.config["scanner"]
+        #         else self.use_default_scanner
+        #     )
+        #     self.maxbotcount = (
+        #         self.config["scanner"]["maxbotcount"]
+        #         if "maxbotcount" in self.config["scanner"]
+        #         else self.maxbotcount
+        #     )
+        #     self.autoscandelay = (
+        #         self.config["scanner"]["autoscandelay"]
+        #         if "autoscandelay" in self.config["scanner"]
+        #         else 0
+        #     )
+        #     self.enable_buy_next = (
+        #         self.config["scanner"]["enable_buy_next"]
+        #         if "enable_buy_next" in self.config["scanner"]
+        #         else True
+        #     )
 
             # self.autostart = (
             #     self.config["scanner"]["autostart"]
@@ -199,18 +200,69 @@ class TelegramBot(TelegramBotBase):
             use_context=True,
         )
 
+        self.helper.load_config()
+
+        # self.handler = TelegramHandler(self.datafolder, self.userid, self.helper)
+        # self.control = TelegramControl(self.datafolder, self.helper)
+        # self.actions = TelegramActions(self.datafolder, self.helper)
+        # self.editor = ConfigEditor(self.datafolder, self.helper)
+        # self.setting = SettingsEditor(self.datafolder, self.helper)
+# 
+#     def reload_config(self):
+#         # Config section for bot pair scanner
+#         self.atr72pcnt = 2.0
+#         self.enableleverage = False
+#         self.use_default_scanner = 1
+#         self.maxbotcount = 0
+#         self.autoscandelay = 0
+#         self.enable_buy_next = True
+#         self.autostart = False
+#         if "scanner" in self.config:
+#             self.atr72pcnt = (
+#                 self.config["scanner"]["atr72_pcnt"]
+#                 if "atr72_pcnt" in self.config["scanner"]
+#                 else self.atr72pcnt
+#             )
+#             self.enableleverage = (
+#                 self.config["scanner"]["enableleverage"]
+#                 if "enableleverage" in self.config["scanner"]
+#                 else self.enableleverage
+#             )
+#             self.use_default_scanner = (
+#                 self.config["scanner"]["use_default_scanner"]
+#                 if "use_default_scanner" in self.config["scanner"]
+#                 else self.use_default_scanner
+#             )
+#             self.maxbotcount = (
+#                 self.config["scanner"]["maxbotcount"]
+#                 if "maxbotcount" in self.config["scanner"]
+#                 else self.maxbotcount
+#             )
+#             self.autoscandelay = (
+#                 self.config["scanner"]["autoscandelay"]
+#                 if "autoscandelay" in self.config["scanner"]
+#                 else 0
+#             )
+#             self.enable_buy_next = (
+#                 self.config["scanner"]["enable_buy_next"]
+#                 if "enable_buy_next" in self.config["scanner"]
+#                 else True
+#             )
+
+        self.helper = TelegramHelper(self.datafolder, self.config, self.config_file)
+
         self.handler = TelegramHandler(self.datafolder, self.userid, self.helper)
         self.control = TelegramControl(self.datafolder, self.helper)
         self.actions = TelegramActions(self.datafolder, self.helper)
         self.editor = ConfigEditor(self.datafolder, self.helper)
+        self.setting = SettingsEditor(self.datafolder, self.helper)
 
     def _question_which_exchange(self, update):
         """start new bot ask which exchange"""
 
         self.exchange = ""
         self.overrides = ""
-
-        update.message.reply_text("Select the exchange:", reply_markup=markup)
+        self.helper.send_telegram_message(update, "Select the exchange:", markup)
 
     def _answer_which_exchange(self, update) -> bool:
         """start bot validate exchange and ask which market/pair"""
@@ -230,8 +282,7 @@ class TelegramBot(TelegramBotBase):
                 self.exchange = "coinbasepro"
         else:
             if self.exchange == "":
-                update.message.reply_text("Invalid Exchange Entered!")
-                # self.newbot_request(update, context)
+                self.helper.send_telegram_message(update, "Invalid Exchange Entered!" )
                 return False
 
         return True
@@ -239,33 +290,28 @@ class TelegramBot(TelegramBotBase):
     def _question_which_pair(self, update):
 
         self.market = ""
-
-        update.message.reply_text(
-            "Which market/pair is this for?", reply_markup=ReplyKeyboardRemove()
-        )
+        self.helper.send_telegram_message(update, "Which market/pair is this for?", ReplyKeyboardRemove())
 
     def _answer_which_pair(self, update) -> bool:
         if update.message.text.lower() == "cancel":
-            update.message.reply_text(
-                "Operation Cancelled", reply_markup=ReplyKeyboardRemove()
+            self.helper.send_telegram_message(
+                update, "Operation Cancelled", ReplyKeyboardRemove()
             )
             return ConversationHandler.END
 
         if self.exchange in ("coinbasepro", "kucoin"):
             p = re.compile(r"^[0-9A-Z]{1,20}\-[1-9A-Z]{2,5}$")
             if not p.match(update.message.text):
-                update.message.reply_text(
-                    "Invalid market format", reply_markup=ReplyKeyboardRemove()
+                self.helper.send_telegram_message(
+                    update, "Invalid market format", ReplyKeyboardRemove()
                 )
-                # self.newbot_exchange(update, context)
                 return False
         elif self.exchange == "binance":
             p = re.compile(r"^[A-Z0-9]{4,25}$")
             if not p.match(update.message.text):
-                update.message.reply_text(
-                    "Invalid market format.", reply_markup=ReplyKeyboardRemove()
+                self.helper.send_telegram_message(
+                    update, "Invalid market format.", ReplyKeyboardRemove()
                 )
-                # self.newbot_exchange(update, context)
                 return False
 
         self.pair = update.message.text
@@ -303,11 +349,7 @@ class TelegramBot(TelegramBotBase):
         ubot = Bot(self.token)
         ubot.set_my_commands(command)
 
-        update.message.reply_text(
-            "<i>Bot Commands Created</i>",
-            parse_mode="HTML",
-            reply_markup=ReplyKeyboardRemove(),
-        )
+        self.helper.send_telegram_message(update, "<i>Bot Commands Created</i>", ReplyKeyboardRemove())
 
     def help(self, update, context):
         """Send a message when the command /help is issued."""
@@ -318,13 +360,13 @@ class TelegramBot(TelegramBotBase):
         )
         helptext += "<b>/margins</b> - <i>show margins for open trade</i>\n"
         helptext += "<b>/trades</b> - <i>show closed trades</i>\n"
-        helptext += "<b>/stats</b> - <i>display stats for market</i>\n"
+        helptext += "<b>/stats</b> - <i>display stats for market</i>\n\n"
         # helptext += "<b>/showinfo</b> - <i>display bot(s) status</i>\n"
         # helptext += "<b>/showconfig</b> - <i>show config for exchange</i>\n\n"
         helptext += "<b>Interactive Command List</b>\n\n"
         helptext += "<b>/controlpanel</b> - <i>show interactive control buttons</i>\n"
         helptext += "<b>/cleandata</b> - <i>check and remove any bad Json files</i>\n"
-        helptext += "<b>/addnew</b> - <i>start the requested pair</i>\n"
+        helptext += "<b>/addnew</b> - <i>start the requested pair</i>\n\n"
         # helptext += "<b>/pausebots</b> - <i>pause all or the selected bot</i>\n"
         # helptext += "<b>/resumebots</b> - <i>resume paused bots</i>\n"
         # helptext += "<b>/stopbots</b> - <i>stop all or the selected bots</i>\n"
@@ -339,38 +381,35 @@ class TelegramBot(TelegramBotBase):
             "<b>/removeexception</b> - <i>remove pair from scanner exception list</i>\n"
         )
 
-        mbot = Telegram(self.token, str(context._chat_id_and_data[0]))
-
-        mbot.send(helptext, parsemode="HTML")
+        self.helper.send_telegram_message(update, helptext)
 
     def trades(self, update, context):
         """List trades"""
-        if not self._checkifallowed(context._user_id_and_data[0], update):
+        if not self._check_if_allowed(context._user_id_and_data[0], update):
             return
 
         self.helper.read_data()
 
         output = ""
-        for time in self.helper.data["trades"]:
+        for dt in self.helper.data["trades"]:
             output = ""
             output = (
-                output + f"<b>{self.helper.data['trades'][time]['pair']}</b>\n{time}"
+                output + f"<b>{self.helper.data['trades'][dt]['pair']}</b>\n{dt}"
             )
             output = (
                 output
-                + f"\n<i>Sold at: {self.helper.data['trades'][time]['price']}   Margin: {self.helper.data['trades'][time]['margin']}</i>\n"
+                + f"\n<i>Sold at: {self.helper.data['trades'][dt]['price']}   Margin: {self.helper.data['trades'][dt]['margin']}</i>\n"
             )
 
             if output != "":
-                mbot = Telegram(self.token, str(context._chat_id_and_data[0]))
-                mbot.send(output, parsemode="HTML")
+                self.helper.send_telegram_message(update, output)
 
     def statsrequest(self, update: Updater, context):
         """Ask which exchange stats are wanted for"""
-        if not self._checkifallowed(context._user_id_and_data[0], update):
+        if not self._check_if_allowed(context._user_id_and_data[0], update):
             return None
 
-        update.message.reply_text("Select the exchange", reply_markup=markup)
+        self.helper.send_telegram_message(update, "Select the exchange", markup)
 
         return CHOOSING
 
@@ -380,9 +419,7 @@ class TelegramBot(TelegramBotBase):
             return None
 
         if update.message.text.lower() == "cancel":
-            update.message.reply_text(
-                "Operation Cancelled", reply_markup=ReplyKeyboardRemove()
-            )
+            self.helper.send_telegram_message(update, "Operation Cancelled", ReplyKeyboardRemove())
             return ConversationHandler.END
 
         if update.message.text in ("Coinbase Pro", "Kucoin", "Binance"):
@@ -391,14 +428,11 @@ class TelegramBot(TelegramBotBase):
                 self.exchange = "coinbasepro"
         else:
             if self.exchange == "":
-                update.message.reply_text("Invalid Exchange Entered!")
+                self.helper.send_telegram_message(update, "Invalid Exchange Entered!")
                 self.statsrequest(update, context)
                 return None
 
-        update.message.reply_text(
-            "Which market/pair do you want stats for?",
-            reply_markup=ReplyKeyboardRemove(),
-        )
+        self.helper.send_telegram_message(update, "Which market/pair do you want stats for?", ReplyKeyboardRemove())
 
         return TYPING_REPLY
 
@@ -408,45 +442,36 @@ class TelegramBot(TelegramBotBase):
             return None
 
         if update.message.text.lower() == "cancel":
-            update.message.reply_text(
-                "Operation Cancelled", reply_markup=ReplyKeyboardRemove()
-            )
+            self.helper.send_telegram_message(update, "Operation Cancelled", ReplyKeyboardRemove())
             return ConversationHandler.END
 
-        if self.exchange == "coinbasepro" or self.exchange == "kucoin":
+        if self.exchange in ("coinbasepro", "kucoin"):
             p = re.compile(r"^[0-9A-Z]{1,20}\-[1-9A-Z]{2,5}$")
             if not p.match(update.message.text):
-                update.message.reply_text(
-                    "Invalid market format", reply_markup=ReplyKeyboardRemove()
-                )
+                self.helper.send_telegram_message(update, "Invalid market format", ReplyKeyboardRemove())
                 self.stats_exchange_received(update, context)
                 return None
         elif self.exchange == "binance":
             p = re.compile(r"^[A-Z0-9]{4,25}$")
             if not p.match(update.message.text):
-                update.message.reply_text(
-                    "Invalid market format.", reply_markup=ReplyKeyboardRemove()
-                )
+                self.helper.send_telegram_message(update, "Invalid market format", ReplyKeyboardRemove())
                 self.stats_exchange_received(update, context)
                 return None
 
         self.pair = update.message.text
 
-        update.message.reply_text(
-            "<i>Gathering Stats, please wait...</i>", parse_mode="HTML"
+        self.helper.send_telegram_message(update, "<i>Gathering Stats, please wait...</i>")
+        
+        output = self.helper.start_process(
+            self.pair, self.exchange, "--stats --live 1", "telegram", True
         )
-
-        output = self.helper.startProcess(
-            self.pair, self.exchange, "--stats", "telegram", True
-        )
-
-        update.message.reply_text(output, parse_mode="HTML")
+        self.helper.send_telegram_message(update, output)
 
         return ConversationHandler.END
 
     def newbot_request(self, update: Updater, context):
         """start new bot ask which exchange"""
-        if not self._checkifallowed(context._user_id_and_data[0], update):
+        if not self._check_if_allowed(context._user_id_and_data[0], update):
             return None
 
         self._question_which_exchange(update)
@@ -455,7 +480,7 @@ class TelegramBot(TelegramBotBase):
 
     def newbot_exchange(self, update, context):
         """start bot validate exchange and ask which market/pair"""
-        if not self._checkifallowed(context._user_id_and_data[0], update):
+        if not self._check_if_allowed(context._user_id_and_data[0], update):
             return None
 
         if not self._answer_which_exchange(update):
@@ -467,7 +492,7 @@ class TelegramBot(TelegramBotBase):
 
     def newbot_any_overrides(self, update, context) -> None:
         """start bot validate market and ask if overrides required"""
-        if not self._checkifallowed(context._user_id_and_data[0], update):
+        if not self._check_if_allowed(context._user_id_and_data[0], update):
             return None
 
         if not self._answer_which_pair(update):
@@ -478,33 +503,32 @@ class TelegramBot(TelegramBotBase):
 
         mark_up = ReplyKeyboardMarkup(r_keyboard, one_time_keyboard=True)
 
-        update.message.reply_text(
-            "Do you want to use any commandline overrides?", reply_markup=mark_up
+        self.helper.send_telegram_message(
+            update, "Do you want to use any commandline overrides?", mark_up
         )
 
         return MARKET
 
     def newbot_market(self, update, context):
         """start bot - ask for overrides if none required ask to save bot"""
-        if not self._checkifallowed(context._user_id_and_data[0], update):
+        if not self._check_if_allowed(context._user_id_and_data[0], update):
             return None
 
         if update.message.text == "No":
             reply_keyboard = [["Yes", "No"]]
-            markup = ReplyKeyboardMarkup(reply_keyboard, one_time_keyboard=True)
-            update.message.reply_text("Do you want to save this?", reply_markup=markup)
+            mark_up = ReplyKeyboardMarkup(reply_keyboard, one_time_keyboard=True)
+            self.helper.send_telegram_message(update, "Do you want to save this?", mark_up)
             return SAVE
 
-        update.message.reply_text(
-            "Tell me any other commandline overrides to use?",
-            reply_markup=ReplyKeyboardRemove(),
+        self.helper.send_telegram_message(
+            update, "Tell me any other commandline overrides to use?", ReplyKeyboardRemove(),
         )
 
         return OVERRIDES
 
     def newbot_overrides(self, update, context):
         """start bot - ask to save bot"""
-        if not self._checkifallowed(context._user_id_and_data[0], update):
+        if not self._check_if_allowed(context._user_id_and_data[0], update):
             return None
 
         # Telegram desktop client can auto replace -- with a single long dash
@@ -514,14 +538,14 @@ class TelegramBot(TelegramBotBase):
         )
 
         reply_keyboard = [["Yes", "No"]]
-        markup = ReplyKeyboardMarkup(reply_keyboard, one_time_keyboard=True)
-        update.message.reply_text("Do you want to save this?", reply_markup=markup)
+        mark_up = ReplyKeyboardMarkup(reply_keyboard, one_time_keyboard=True)
+        self.helper.send_telegram_message(update, "Do you want to save this?", mark_up)
 
         return SAVE
 
     def newbot_save(self, update, context):
         """start bot - save if required ask if want to start"""
-        if not self._checkifallowed(context._user_id_and_data[0], update):
+        if not self._check_if_allowed(context._user_id_and_data[0], update):
             return None
 
         if update.message.text == "Yes":
@@ -536,9 +560,9 @@ class TelegramBot(TelegramBotBase):
                         }
                     )
                     self.helper.write_data()
-                    update.message.reply_text(f"{self.pair} saved \u2705")
+                    self.helper.send_telegram_message(update, f"{self.pair} saved \u2705")
                 else:
-                    update.message.reply_text(
+                    self.helper.send_telegram_message(update,
                         f"{self.pair} already setup, no changes made."
                     )
             else:
@@ -551,44 +575,44 @@ class TelegramBot(TelegramBotBase):
                     }
                 )
                 self.helper.write_data()
-                update.message.reply_text(f"{self.pair} saved")
+                self.helper.send_telegram_message(update, f"{self.pair} saved")
 
         reply_keyboard = [["Yes", "No"]]
-        markup = ReplyKeyboardMarkup(reply_keyboard, one_time_keyboard=True)
-        update.message.reply_text("Do you want to start this bot?", reply_markup=markup)
+        mark_up = ReplyKeyboardMarkup(reply_keyboard, one_time_keyboard=True)
+        self.helper.send_telegram_message(update, "Do you want to start this bot?", mark_up)
 
         return START
 
     def newbot_start(self, update, context, startmethod: str = "telegram") -> None:
         """start bot - start bot if want"""
-        if not self._checkifallowed(context._user_id_and_data[0], update):
+        if not self._check_if_allowed(context._user_id_and_data[0], update):
             return None
 
         if update.message.text == "No":
-            update.message.reply_text(
-                "Command Complete, have a nice day.", reply_markup=ReplyKeyboardRemove()
+            self.helper.send_telegram_message(update,
+                "Command Complete, have a nice day.", ReplyKeyboardRemove()
             )
             return ConversationHandler.END
 
         if (
-            self.helper.startProcess(
+            self.helper.start_process(
                 self.pair, self.exchange, self.overrides, startmethod
             )
-            == False
+            is False
         ):
-            update.message.reply_text(
+            self.helper.send_telegram_message(update,
                 f"{self.pair} is already running, no action taken.",
-                reply_markup=ReplyKeyboardRemove(),
+                ReplyKeyboardRemove(),
             )
         else:
             if startmethod != "scanner":
-                update.message.reply_text(
+                self.helper.send_telegram_message(update,
                     f"{self.pair} crypto bot Starting",
-                    reply_markup=ReplyKeyboardRemove(),
+                    ReplyKeyboardRemove(),
                 )
 
-        update.message.reply_text(
-            "Command Complete, have a nice day.", reply_markup=ReplyKeyboardRemove()
+        self.helper.send_telegram_message(update,
+            "Command Complete, have a nice day.", ReplyKeyboardRemove()
         )
 
         return ConversationHandler.END
@@ -623,7 +647,7 @@ class TelegramBot(TelegramBotBase):
             return False
 
     def _cleandata(self):
-        jsonfiles = self.helper.getActiveBotList()
+        jsonfiles = self.helper.get_active_bot_list()
         for i in range(len(jsonfiles), 0, -1):
             jfile = jsonfiles[i - 1]
 
@@ -666,7 +690,7 @@ class TelegramBot(TelegramBotBase):
         return EXCEPT_EXCHANGE
 
     def ExceptionPair(self, update, context):
-        if not self._checkifallowed(context._user_id_and_data[0], update):
+        if not self._check_if_allowed(context._user_id_and_data[0], update):
             return
 
         self._answer_which_exchange(update)
@@ -677,7 +701,7 @@ class TelegramBot(TelegramBotBase):
 
     def ExceptionAdd(self, update, context):
         """start bot - save if required ask if want to start"""
-        if not self._checkifallowed(context._user_id_and_data[0], update):
+        if not self._check_if_allowed(context._user_id_and_data[0], update):
             return None
 
         self._answer_which_pair(update)
@@ -690,136 +714,137 @@ class TelegramBot(TelegramBotBase):
         if not self.pair in self.helper.data["scannerexceptions"]:
             self.helper.data["scannerexceptions"].update({self.pair: {}})
             self.helper.write_data()
-            update.message.reply_text(
+            self.helper.send_telegram_message(update,
                 f"{self.pair} Added to Scanner Exception List \u2705",
-                reply_markup=ReplyKeyboardRemove(),
+                ReplyKeyboardRemove(),
             )
         else:
-            update.message.reply_text(
+            self.helper.send_telegram_message(update,
                 f"{self.pair} Already on exception list",
-                reply_markup=ReplyKeyboardRemove(),
+                ReplyKeyboardRemove(),
             )
 
         return ConversationHandler.END
 
     def ExceptionRemove(self, update, context):
-        if not self._checkifallowed(context._user_id_and_data[0], update):
+        if not self._check_if_allowed(context._user_id_and_data[0], update):
             return
 
-        self.control.askExceptionBotList(update)
+        self.control.ask_exception_bot_list(update)
         return
 
     def marginrequest(self, update, context):
-        if not self._checkifallowed(context._user_id_and_data[0], update):
+        if not self._check_if_allowed(context._user_id_and_data[0], update):
             return
 
-        self.handler.askMarginType(update)
+        self.handler.ask_margin_type(update)
         return
 
     def showbotinfo(self, update, context) -> None:
         """Show running bot status"""
-        if not self._checkifallowed(context._user_id_and_data[0], update):
+        if not self._check_if_allowed(context._user_id_and_data[0], update):
             return
 
-        self.actions.getBotInfo(update)
+        self.actions.get_bot_info(update)
         return
 
     def sellrequest(self, update, context):
         """Manual sell request (asks which coin to sell)"""
-        self.control.askSellBotList(update)
+        self.control.ask_sell_bot_list(update)
         return
 
     def buyrequest(self, update, context):
         """Manual buy request (asks which coin to buy)"""
-        if not self._checkifallowed(context._user_id_and_data[0], update):
+        if not self._check_if_allowed(context._user_id_and_data[0], update):
             return
 
-        self.control.askBuyBotList(update)
+        self.control.ask_buy_bot_list(update)
         return
 
     def showconfigrequest(self, update, context):
         """display config settings (ask which exchange)"""
-        if not self._checkifallowed(context._user_id_and_data[0], update):
+        if not self._check_if_allowed(context._user_id_and_data[0], update):
             return
 
-        self.handler.askConfigOptions(update)
+        self.handler.ask_config_options(update)
         return
 
     def pausebotrequest(self, update, context) -> None:
         """Ask which bots to pause"""
-        if not self._checkifallowed(context._user_id_and_data[0], update):
+        if not self._check_if_allowed(context._user_id_and_data[0], update):
             return
 
-        self.control.askPauseBotList(update)
+        self.control.ask_pause_bot_list(update)
 
     def restartbotrequest(self, update, context) -> None:
         """Ask which bot to restart"""
-        if not self._checkifallowed(context._user_id_and_data[0], update):
+        if not self._check_if_allowed(context._user_id_and_data[0], update):
             return
 
-        self.control.askResumeBotList(update)
+        self.control.ask_resume_bot_list(update)
 
     def startallbotsrequest(self, update, context) -> None:
         """Ask which bot to start from start list (or all)"""
-        if not self._checkifallowed(context._user_id_and_data[0], update):
+        if not self._check_if_allowed(context._user_id_and_data[0], update):
             return
 
-        self.control.askStartBotList(update)
+        self.control.ask_start_bot_list(update)
         return
 
     def stopbotrequest(self, update, context) -> None:
         """ask which active bots to stop (or all)"""
-        if not self._checkifallowed(context._user_id_and_data[0], update):
+        if not self._check_if_allowed(context._user_id_and_data[0], update):
             return
 
-        self.control.askStopBotList(update)
+        self.control.ask_stop_bot_list(update)
         return
 
     def deleterequest(self, update, context):
         """ask which bot to delete"""
-        if not self._checkifallowed(context._user_id_and_data[0], update):
+        if not self._check_if_allowed(context._user_id_and_data[0], update):
             return
 
-        self.control.askDeleteBotList(update)
+        self.control.ask_delete_bot_list(update)
 
     def StartScanning(self, update, context):
-        if not self._checkifallowed(context._user_id_and_data[0], update):
+        if not self._check_if_allowed(context._user_id_and_data[0], update):
             return
 
-        self.handler._checkScheduledJob(update)
-        logger.info("Start scanning using default scanner? %s", True if self.use_default_scanner == 1 else False)
-        self.actions.StartMarketScan(
+        self.handler._check_scheduled_job(update)
+        logger.info("Start scanning using default scanner? %s", bool(self.helper.use_default_scanner))
+        self.helper.send_telegram_message(update, "Operation Started")
+        self.actions.start_market_scan(
             update,
-            True if self.use_default_scanner == 1 else False,
+            True if self.helper.use_default_scanner == 1 else False,
             False if len(context.args) > 0 and context.args[0] == "debug" else True,
             False if len(context.args) > 0 and context.args[0] == "noscan" else True
         )
 
     def StopScanning(self, update, context):
-        if not self._checkifallowed(context._user_id_and_data[0], update):
+        if not self._check_if_allowed(context._user_id_and_data[0], update):
             return
-        self.handler._removeScheduledJob(update)
+        self.handler._remove_scheduled_job(update)
 
     def cleandata(self, update, context) -> None:
-        if not self._checkifallowed(context._user_id_and_data[0], update):
+        if not self._check_if_allowed(context._user_id_and_data[0], update):
             return
 
         self._cleandata()
 
-        self.actions.getBotInfo(update)
-        update.message.reply_text("Operation Complete")
+        self.actions.get_bot_info(update)
+        self.helper.send_telegram_message(update, "Operation Complete")
 
     def RestartBots(self, update, context):
-        if not self._checkifallowed(context._user_id_and_data[0], update):
+        if not self._check_if_allowed(context._user_id_and_data[0], update):
             return None
 
-        self.control.askRestartBotList(update)
+        self.control.ask_restart_bot_list(update)
 
     def StartOpenOrderBots(self, update, context):
-        if not self._checkifallowed(context._user_id_and_data[0], update):
+        if not self._check_if_allowed(context._user_id_and_data[0], update):
             return None
 
-        self.actions.startOpenOrders(update)
+        self.actions.start_open_orders(update)
 
     def statstwo(self, update, context):
 
@@ -834,8 +859,8 @@ class TelegramBot(TelegramBotBase):
                 if file.__contains__("kucoin"):
                     exchange = "kucoin"
 
-                update.message.reply_text(
-                    "<i>Gathering Stats, please wait...</i>", parse_mode="HTML"
+                self.helper.send_telegram_message(update,
+                    "<i>Gathering Stats, please wait...</i>"
                 )
 
                 with open(
@@ -857,12 +882,12 @@ class TelegramBot(TelegramBotBase):
                 output = subprocess.getoutput(
                     f"python3 pycryptobot.py --stats --exchange {exchange}  --statgroup {pairs}  "
                 )
-                update.message.reply_text(output, parse_mode="HTML")
+                self.helper.send_telegram_message(update, output)
                 sleep(30)
-                update.message.reply_text("Pausing before next set", parse_mode="HTML")
+                self.helper.send_telegram_message(update, "Pausing before next set")
 
     def getBotList(self, update, context):
-        if not self._checkifallowed(context._user_id_and_data[0], update):
+        if not self._check_if_allowed(context._user_id_and_data[0], update):
             return None
 
         query = update.callback_query
@@ -873,7 +898,7 @@ class TelegramBot(TelegramBotBase):
 
         buttons = []
 
-        for market in self.helper.getActiveBotList("all"):
+        for market in self.helper.get_active_bot_list("active"):
             while self.helper.read_data(market) == False:
                 sleep(0.2)
 
@@ -881,19 +906,9 @@ class TelegramBot(TelegramBotBase):
                 buttons.append(InlineKeyboardButton(market, callback_data=f"bot_{market}"))
 
         if len(buttons) > 0:
-            try:
-                query.edit_message_text("<b>Select a market</b>",
-                    reply_markup=self.control._sortInlineButtons(buttons, "bot"),
-                    parse_mode="HTML")
-            except:
-                update.effective_message.reply_html(
-                    "<b>Select a market</b>",
-                    reply_markup=self.control._sortInlineButtons(buttons, "bot"))
+            self.helper.send_telegram_message(update, "<b>Select a market</b>", self.control.sort_inline_buttons(buttons, "bot"))
         else:
-            try:
-                query.edit_message_text("<b>No bots found.</b>", parse_mode="HTML")
-            except:
-                update.effective_message.reply_html(f"<b>No bots found.</b>")
+            self.helper.send_telegram_message(update, "<b>No bots found.</b>")
 
     #     def UpdateBuyMaxSize(self, update, context):
     #
@@ -916,12 +931,13 @@ class TelegramBot(TelegramBotBase):
 
         userid = context._user_id_and_data[0]
 
-        if self._checkifallowed(userid, update):
-            key_markup = self.handler.getRequest()
-            update.message.reply_text(
+        if self._check_if_allowed(userid, update):
+            self.helper.load_config()
+            key_markup = self.handler.get_request()
+            self.helper.send_telegram_message(update,
                 "<b>PyCryptoBot Command Panel.</b>",
-                reply_markup=key_markup,
-                parse_mode="HTML",
+                key_markup,
+                context
             )
 
     # def ExitBot(self, update, context):
@@ -978,7 +994,7 @@ def main():
 
     dp.add_handler(CommandHandler("statsgroup", botconfig.statstwo))
     # Response to Question handler
-    dp.add_handler(CallbackQueryHandler(botconfig.handler.getResponse))
+    dp.add_handler(CallbackQueryHandler(botconfig.handler.get_response))
 
     dp.add_handler(CommandHandler("controlPanel", botconfig.Request))