+++ CHANGELOG.md
@@ -8,6 +8,12 @@ Upgrade version:
 Upgrade library dependencies (if required):
 - python3 -m pip install -r requirements.txt -U
 
+## [5.2.0] - 2022-01-11
+
+- updated telegram bot config editor
+- added option to enable/disable screener/scanner notifications
+- update to default parser to set to correct object type
+
 ## [5.1.3] - 2022-01-03
 
 - fixed spelling mistake in comment